requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
requests==2.31.0
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.6.1
